pyproject.toml
[project]
name = "assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "elevenlabs>=2.9.2",
    "fastapi>=0.116.1",
    "openai>=1.99.6",
    "python-dotenv>=1.1.1",
    "sarvamai>=0.1.18",
    "uvicorn>=0.35.0",
]
